refactory: Modify test files
