+++ cover.out
@@ -1,48 +1,14 @@
-mode: count
-github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:16.74,27.16 4 1
-github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:27.16,29.3 1 0
-github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:31.2,31.18 1 1
-github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:34.24,36.2 1 1
-github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:38.32,40.18 1 3
-github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:40.18,42.3 1 1
-github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:44.2,44.22 1 2
-github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:44.22,46.3 1 0
-github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:48.2,48.19 1 2
-github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:48.19,50.3 1 0
-github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:52.2,52.12 1 2
-github.com/reangeline/foodscan_backend/pkg/entities/id.go:7.17,9.2 1 1
-github.com/reangeline/foodscan_backend/pkg/entities/id.go:11.36,14.2 2 0
+mode: set
 github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:16.74,27.16 4 1
 github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:27.16,29.3 1 0
 github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:31.2,31.18 1 1
 github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:34.24,36.2 1 1
 github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:38.32,40.18 1 1
-github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:40.18,42.3 1 0
+github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:40.18,42.3 1 1
 github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:44.2,44.22 1 1
 github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:44.22,46.3 1 0
 github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:48.2,48.19 1 1
 github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:48.19,50.3 1 0
 github.com/reangeline/foodscan_backend/internal/domain/entities/user.go:52.2,52.12 1 1
-github.com/reangeline/foodscan_backend/internal/domain/usecases/user_usecase.go:16.87,18.2 1 1
-github.com/reangeline/foodscan_backend/internal/domain/usecases/user_usecase.go:24.90,28.13 2 2
-github.com/reangeline/foodscan_backend/internal/domain/usecases/user_usecase.go:28.13,30.3 1 1
-github.com/reangeline/foodscan_backend/internal/domain/usecases/user_usecase.go:32.2,33.16 2 1
-github.com/reangeline/foodscan_backend/internal/domain/usecases/user_usecase.go:33.16,35.3 1 0
-github.com/reangeline/foodscan_backend/internal/domain/usecases/user_usecase.go:37.2,37.63 1 1
-github.com/reangeline/foodscan_backend/internal/domain/usecases/user_usecase.go:37.63,39.3 1 0
-github.com/reangeline/foodscan_backend/internal/domain/usecases/user_usecase.go:41.2,41.12 1 1
-github.com/reangeline/foodscan_backend/internal/domain/usecases/user_usecase.go:44.68,46.16 2 2
-github.com/reangeline/foodscan_backend/internal/domain/usecases/user_usecase.go:46.16,48.3 1 1
-github.com/reangeline/foodscan_backend/internal/domain/usecases/user_usecase.go:50.2,50.27 1 1
-github.com/reangeline/foodscan_backend/internal/domain/usecases/user_usecase.go:50.27,52.3 1 0
-github.com/reangeline/foodscan_backend/internal/domain/usecases/user_usecase.go:54.2,54.18 1 1
-github.com/reangeline/foodscan_backend/internal/domain/usecases/user_usecase.go:57.78,60.16 2 1
-github.com/reangeline/foodscan_backend/internal/domain/usecases/user_usecase.go:60.16,62.3 1 0
-github.com/reangeline/foodscan_backend/internal/domain/usecases/user_usecase.go:64.2,64.18 1 1
 github.com/reangeline/foodscan_backend/pkg/entities/id.go:7.17,9.2 1 1
 github.com/reangeline/foodscan_backend/pkg/entities/id.go:11.36,14.2 2 0
-github.com/reangeline/foodscan_backend/tests/unit/internal/infra/database/user_repository_memory.go:15.54,19.2 1 1
-github.com/reangeline/foodscan_backend/tests/unit/internal/infra/database/user_repository_memory.go:21.91,24.2 2 1
-github.com/reangeline/foodscan_backend/tests/unit/internal/infra/database/user_repository_memory.go:26.87,29.9 2 3
-github.com/reangeline/foodscan_backend/tests/unit/internal/infra/database/user_repository_memory.go:29.9,31.3 1 1
-github.com/reangeline/foodscan_backend/tests/unit/internal/infra/database/user_repository_memory.go:33.2,38.8 1 2
