+++ go.mod
@@ -5,7 +5,6 @@ go 1.20
 require (
 	github.com/99designs/gqlgen v0.17.32
 	github.com/go-chi/chi/v5 v5.0.8
-	github.com/golang/mock v1.4.4
 	github.com/google/uuid v1.1.2
 	github.com/google/wire v0.5.0
 	github.com/lib/pq v1.10.9
