+++ internal/domain/entities/user_test.go
